fix size count and add comments

+++ done/pictDBM/db_list.c
@@ -10,8 +10,9 @@
 
 void do_list (const struct pictdb_file file) {
     print_header(file.header);
-    int count = 0;
+    size_t count = 0;
     for (size_t i = 0; i < MAX_MAX_FILES; ++i) {
+        //
         if (file.metadata[i].is_valid == NON_EMPTY) {
             print_metadata(file.metadata[i]);
             count += 1;
@@ -20,5 +21,4 @@ void do_list (const struct pictdb_file file) {
     if (count == 0) {
         printf("<< empty database >>\n");
     }
-    
 }
